requirements.txt
alembic==1.12.1
aniso8601==10.0.0
bcrypt==4.2.1
blinker==1.8.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
filelock==3.16.1
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fsspec==2025.3.0
greenlet==3.1.1
gunicorn==23.0.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==2.1.5
packaging==24.2
ps==0.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.23
SQLAlchemy-serializer==1.4.12
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==2.3.7
zipp==3.20.2
